bot/requirements.txt
requests
python-dotenv
pytz
cachetools
simplejson
pytest
nextcord
pymongo